Refine Apple Pay iOS callback routing
